chore: rename scrvusd queries to match naming convention of other queries
