+++ apps/main/src/loan/components/PageCrvUsdStaking/DepositWithdraw/DepositModule.tsx
@@ -2,7 +2,7 @@ import BigNumber from 'bignumber.js'
 import { RCCrvUSDLogoXS, RCScrvUSDLogoXS } from 'ui/src/images'
 import { useAccount } from 'wagmi'
 import { isLoading } from '@/loan/components/PageCrvUsdStaking/utils'
-import { useScrvUsdUserBalances } from '@/loan/entities/scrvusdUserBalances'
+import { useScrvUsdUserBalances } from '@/loan/entities/scrvusd-userBalances'
 import useStore from '@/loan/store/useStore'
 import Box from '@ui/Box'
 import { t } from '@ui-kit/lib/i18n'
