+++ apps/main/src/loan/components/PageCrvUsdStaking/Statistics/RevenueChartTooltip.tsx
@@ -2,7 +2,7 @@ import { TooltipProps } from 'recharts'
 import type { NameType, ValueType } from 'recharts/types/component/DefaultTooltipContent'
 import LegendLine from '@/loan/components/PageCrvUsdStaking/Statistics/components/LegendLine'
 import { TOOLTIP_MAX_WIDTH, TOOLTIP_MAX_WIDTH_MOBILE } from '@/loan/components/PageCrvUsdStaking/Statistics/constants'
-import type { ScrvUsdYieldWithAverages } from '@/loan/entities/scrvusdYield'
+import type { ScrvUsdYieldWithAverages } from '@/loan/entities/scrvusd-yield'
 import { Paper, Stack, Typography } from '@mui/material'
 import { useTheme } from '@mui/material/styles'
 import { formatDate } from '@ui/utils'
