+++ apps/main/src/loan/components/PageCrvUsdStaking/Statistics/StatsStack.tsx
@@ -1,6 +1,6 @@
-import { useScrvUsdRevenue } from '@/loan/entities/scrvusdRevenue'
-import { useScrvUsdStatistics } from '@/loan/entities/scrvusdStatistics'
-import { useScrvUsdYield } from '@/loan/entities/scrvusdYield'
+import { useScrvUsdRevenue } from '@/loan/entities/scrvusd-revenue'
+import { useScrvUsdStatistics } from '@/loan/entities/scrvusd-statistics'
+import { useScrvUsdYield } from '@/loan/entities/scrvusd-yield'
 import Grid from '@mui/material/Grid'
 import { t } from '@ui-kit/lib/i18n'
 import type { TimeOption } from '@ui-kit/lib/types/scrvusd'
