+++ apps/main/src/loan/components/PageCrvUsdStaking/UserPosition/index.tsx
@@ -2,8 +2,8 @@ import BigNumber from 'bignumber.js'
 import { useAccount } from 'wagmi'
 import { isReady } from '@/loan/components/PageCrvUsdStaking/utils'
 import { oneMonthProjectionYield, oneYearProjectionYield } from '@/loan/components/PageCrvUsdStaking/utils'
-import { useScrvUsdStatistics } from '@/loan/entities/scrvusdStatistics'
-import { useScrvUsdUserBalances } from '@/loan/entities/scrvusdUserBalances'
+import { useScrvUsdStatistics } from '@/loan/entities/scrvusd-statistics'
+import { useScrvUsdUserBalances } from '@/loan/entities/scrvusd-userBalances'
 import useStore from '@/loan/store/useStore'
 import { Card, CardHeader, Stack } from '@mui/material'
 import Grid from '@mui/material/Grid'
