+++ apps/main/src/loan/components/PageCrvUsdStaking/index.tsx
@@ -6,7 +6,7 @@ import Statistics from '@/loan/components/PageCrvUsdStaking/Statistics'
 import StatsBanner from '@/loan/components/PageCrvUsdStaking/StatsBanner'
 import UserInformation from '@/loan/components/PageCrvUsdStaking/UserInformation'
 import UserPosition from '@/loan/components/PageCrvUsdStaking/UserPosition'
-import { useScrvUsdUserBalances } from '@/loan/entities/scrvusdUserBalances'
+import { useScrvUsdUserBalances } from '@/loan/entities/scrvusd-userBalances'
 import useStore from '@/loan/store/useStore'
 import type { NetworkUrlParams } from '@/loan/types/loan.types'
 import { Stack, useMediaQuery } from '@mui/material'
