+++ apps/main/src/loan/entities/scrvusd-revenue.ts
