+++ apps/main/src/loan/entities/scrvusd-statistics.ts
