+++ apps/main/src/loan/entities/scrvusd-userStats.ts
