+++ apps/main/src/loan/entities/scrvusd-yield.ts
