+++ apps/main/src/loan/store/createScrvUsdSlice.ts
@@ -3,8 +3,8 @@ import lodash from 'lodash'
 import type { GetState, SetState } from 'zustand'
 import type { DepositWithdrawModule, StatisticsChart } from '@/loan/components/PageCrvUsdStaking/types'
 import { SCRVUSD_GAS_ESTIMATE } from '@/loan/constants'
-import type { ScrvUsdUserBalances } from '@/loan/entities/scrvusdUserBalances'
-import { invalidateScrvUsdUserBalances } from '@/loan/entities/scrvusdUserBalances'
+import type { ScrvUsdUserBalances } from '@/loan/entities/scrvusd-userBalances'
+import { invalidateScrvUsdUserBalances } from '@/loan/entities/scrvusd-userBalances'
 import networks from '@/loan/networks'
 import type { State } from '@/loan/store/useStore'
 import { type ChainId, FetchStatus, TransactionStatus } from '@/loan/types/loan.types'
